商品削除機能の修正
#what
商品削除機能の修正
#why
出品者のみが削除できるように変更

+++ app/controllers/items_controller.rb
@@ -39,8 +39,10 @@ class ItemsController < ApplicationController
 
   def destroy
     item = Item.find(params[:id])
-    item.destroy
-    redirect_to root_path
+    unless item.user == current_user
+      item.destroy
+      redirect_to root_path
+    end
   end
